+++ R/tsxtreme-package.R
@@ -37,24 +37,24 @@
 #' [thetaruns()] and some basic functions handling the Laplace
 #' distribution are also available in [dlapl()].
 #' 
-#' @references
-#' Coles, S., Heffernan, J. E. and Tawn, J. A. (1999) Dependence measures for
-#' extreme value analyses. _Extremes_, **2**, 339--365.
+#' Read `vignette("time-series-extremes")` to get more details and a gentle
+#' introduction to the functionality of the package.
 #' 
-#' Davison, A. C. and Smith, R. L. (1990) Models for exceedances over high
-#' thresholds. _Journal of the Royal Statistical Society Series B_,
-#' **52**, 393--442.
+#' @references
+#' Lugrin, T., Davison, A. C. and Tawn, J. A. (2016) Bayesian uncertainty
+#' management in temporal dependence of extremes. _Extremes_, **19**, 491--515.
 #' 
 #' Heffernan, J. E. and Tawn, J. A. (2004) A conditional approach for
 #' multivariate extreme values.
 #' _Journal of the Royal Statistical Society Series B_, **66**, 497--546.
 #' 
-#' Ledford, W. A. and Tawn, J. A. (2003) Diagnostics for dependence within
-#' time series extremes. _Journal of the Royal Statistical Society Series B_,
-#' **65**, 521--543.
+#' Keef, C., Papastathopoulos, I. and Tawn, J. A. (2013) Estimation of the
+#' conditional distribution of a multivariate variable given that one of its
+#' components is large: Additional constraints for the Heffernan and Tawn model.
+#' _Journal of Multivariate Analysis_, **115**, 396–404.
 #' 
-#' Lugrin, T., Davison, A. C. and Tawn, J. A. (2016) Bayesian uncertainty
-#' management in temporal dependence of extremes. _Extremes_, **19**, 491--515.
+#' Coles, S., Heffernan, J. E. and Tawn, J. A. (1999) Dependence measures for
+#' extreme value analyses. _Extremes_, **2**, 339--365. 
 #' 
 #' @seealso [thetafit()], [chifit()], [depfit()]
 #' @keywords internal